Basic support for serial ports

+++ include/io.h
@@ -1,10 +1,12 @@
 #ifndef IO_H
 #define IO_H
 
+#include "definitions.h"
+
 unsigned char inb(uint32_t port);
 unsigned short inw(uint32_t port);
 
-void outb(uint32_t port, int32_t value);
-void outw(uint32_t port, int32_t value);
+void outb(uint32_t port, uint8_t value);
+void outw(uint32_t port, uint16_t value);
 
 #endif
